Improve file naming consistency
- Rename generic filenames (credit.py) to match variable names
- Split income.py into separate files for total_disability_payments and section_22_income
- Fix Illinois filename typos and remove space from il_cta_reduced_fare_program.py
- Follows one-variable-per-file standard

Salvages improvements from old PR #1276
