+++ policyengine_us/variables/gov/irs/credits/elderly_and_disabled/section_22_income.py
@@ -1,15 +1,6 @@
 from policyengine_us.model_api import *
 
 
-class total_disability_payments(Variable):
-    value_type = float
-    entity = Person
-    label = "Disability (total) payments"
-    unit = USD
-    documentation = "Wages (or payments in lieu thereof) paid to an individual for permanent and total disability"
-    definition_period = YEAR
-
-
 class section_22_income(Variable):
     value_type = float
     entity = TaxUnit
